Drop support for Go 1.21 in dice example (#5800)
Resolves #5359
